Extract command palette surface
