+++ docs/DECISIONS.md
@@ -159,6 +159,7 @@
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
 - MCP extension surface state lives in `QuillCodeMCPSurface.swift`, MCP tool/resource/prompt JSON request parsing lives in `WorkspaceMCPRequests.swift`, MCP subprocess lifecycle and session routing live in `WorkspaceMCPRuntime.swift`, concrete stdio launch/prober construction lives behind `WorkspaceMCPServerLaunching` in `WorkspaceMCPServerLauncher.swift`, and Ready-server dynamic tool catalog generation lives in `WorkspaceMCPToolCatalog.swift`. `WorkspaceModel` owns selected-project manifest lookup, UI notices, top-bar status, and state application; surface labels, probe-summary compatibility, request aliases, nested-argument normalization, process/session launch seams, process handles, and catalog text stay in focused helpers with parity gates.
 - Workspace command IDs reduce through `WorkspaceCommandPlan` before touching `WorkspaceModel` side effects. The planner owns prefix parsing, draft-prefill mapping, static action mapping, automation quick recurrence parsing, and canonical command-ID-to-tool-name mapping, while `WorkspaceModel` owns mutations and tool dispatch. This keeps command palette rows, slash-command templates, automation row actions, MCP lifecycle actions, memory actions, and git shortcuts from drifting as the command set grows.
+- Command palette surface behavior lives in `WorkspaceCommandPaletteSurface.swift`. Command records, top-bar overflow projection, automation and Computer Use command factories, category ordering, slash/action query scoping, and ranking/scoring stay beside the command contract instead of inside the aggregate `WorkspaceSurface.swift` payload.
 - Automation record construction and due-run draft creation live in `WorkspaceAutomationEngine.swift`, not inside `WorkspaceModel`. `WorkspaceAutomationFactory` owns concrete `QuillAutomation` records and schedule helpers, `WorkspaceAutomationRunner` owns due filtering, recurrence advancement, and generated thread drafts, and `WorkspaceModel` applies those drafts to UI selection, project refresh, stores, and notification reports. This keeps scheduler behavior testable without pulling project persistence or top-bar state into pure automation tests.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
 - The first memories slice is auditable and user-authored. QuillCode loads bounded `.md`, `.txt`, and `.json` notes from `~/.quillcode/memories` and project `.quillcode/memories`, rejects symlinks and paths outside the memory root, truncates oversized notes with visible labels, snapshots loaded memories onto threads, injects them into TrustedRouter as non-command background context, and surfaces the count/sources through the top bar, Memories pane, static HTML renderer, Playwright harness, slash command, and macOS menu bar. Users can explicitly save global memories with `/remember text`, the Add memory command-palette action, or a model-authored `host.memory.remember` tool call when they ask QuillCode to remember a stable preference/fact; saved notes create transcript/tool audit records, refresh memory context immediately, and reject obvious credentials, tokens, passwords, and private keys. Global memories can be forgotten from the Memories pane and produce a transcript notice. Project memories stay read-only in-app because they are project-owned files. Richer redaction review, idle Chronicle jobs, editing UI, conflict resolution, and fully autonomous inference of memories without explicit user intent are separate safety/UX milestones.
